backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
pydantic>=2.6
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
redis
